extended port to support streaming of samples and tag_t
Renamed the reader/writer in the public API to more explicitly identify if the buffer is intended to transmit samples of type <T> or tags of type <tag_t> (N.B. functionally these buffers are otherwise identical).

This implementation assumed the same buffer size for streams and tags since -- in theory -- one could attach as many tags as samples per processing iteration.

N.B. for large buffers and large numbers of ports/blocks in the application this can be wasteful and one should find a good policy w.r.t. the usually rare occurrence of tags compared to samples. I.e. 10%|1%|0.1% of buffer size being allocated for tags?!?!

Signed-off-by: Ralph J. Steinhagen <r.steinhagen@gsi.de>