Fix relocating Port instances (#235)
* Convert qa_DynamicBlock to a real test

Previously the test only printed the result with the cout_sink, which is
fine for manual testing. Convert this to a real boost unit test that
actually asserts some stuff, so that this can be automatically tested.

* Add a test for dynamically reallocating the ports collection

This test will fail when we switch from std::list to std::vector.

This also adapts the simulated scheduler control flow so that each time
only at most one work is done, as this allows us to reason much more
easily about the computed values.

* Fix reallocating Port instances

The move constructor was accidentally missing some fields, which caused
all connections to break on reallocation.
This is the reason why the qa_DynamicBlock test could not use an
std::vector to hold the input ports and instead worked around the
problem by using a std::list.

With the source of the problem fixed now, we switch the test back to
using a std::vector.

Note that this is not the final end of the story. The relationship
between DynamicPort and Port is still a bit of a complicated one,
because the lifetime of these is so tightly coupled together.
DynamicPort objects store a reference to their Port counterpart, so
reallocating a Port without informing (or recreating) a DynamicPort will
quickly lead to problems.

A possible way to fix this in the future is to add a callback to Port,
that will be called on its move constructor to allow the DynamicPort to
adapt its reference to Port.

Fixes https://github.com/fair-acc/opendigitizer/issues/90

* multi_adder: Consume inputs outside of an assert()

This one was the kind of bug that makes any programmer have nightmares:
The test failed only on release builds, it worked fine on Debug builds,
where you would want to actually debug the problem. Looks like
printf-style debugging is back on the menu.

Turns out that an essential line was wrapped in an assert() which of
course is completely disabled in release builds.