Frank/datasink metadata (#283)
* DataSinkRegistry: Make lookup by signal_name thread-safe

As signal_name can change during execution, it cannot be accessed
from other threads. Thus the initial registration must happen after
initialisation, when the initial settings were applied.

For that, make use of the StateMachine's init() callback, and rename
the old init() to performInit() to avoid naming collisions.

* DataSink: Fix signal metadata propagation

Use the automatic propagation of signal_name, signal_unit etc.
properties via default tags instead of trying to handle it manually,
which didn't work correctly with the automatism.

* Fix metadata forwarding when the graph is loaded from GRC

Make sure to add the block to the graph with the properties set, so
that the initial tag containing signal_name, sample_rate, etc. is
created.

* Use properly formed messages for scheduler state notifications

Make sure the messages follow the intended pattern, using the Kind/Data
keys.

* Do not register block at init time

Revert the renaming of Block::init() to Block::performInit().

Instead of registering the DataSink at init time, do so in start(),
and unregister in stop().

This needs changes in the tests, as a poller/callback can now only
be registered while the execution is already running, the
pollers/callbacks might not get all the data from the start.
Add a initial delay (new block testing::Delay) to work around that.