Update vir simd 0.4 (#404)
* Fix qa_Converter when targeting AVX-512

Only convert simd types that are supported by the implementation.

* CMake: Improve CMAKE_BUILD_TYPE usage and docs

* Replace meta::simdize by vir::simdize; adjust implementation accordingly

* Bump to vir-simd v0.4.0
* Replace gr::meta::simdize with vir::simdize.
* Replace gr::meta::simdize_size_v with ::size() on simdized types.
* "Document" that vir::transform can be used, where std::transform and
std::ranges::transform are documented.

* Remove all offsets from processOne(_simd)

Remove:
- invokeProcessOneWithOrWithoutOffset
- exact_argument_type
- can_processOne_with_offset_invoke_test
- can_processOne_simd_with_offset
- can_processOne_scalar_with_offset
- can_processOne_with_offset

Modify MergedGraph apply_left, apply_right, processOne, and
processOne_simd to not require an offset argument anymore.

Modify all calls to processOne(_simd) to never pass an offset.

Modify TagMonitors processOne to generateTag without offset. The
published tag is now unconditional 0 all the time.

* Fix non-const processOne signatures to not accept simd arguments

Either the function is const and can work with chunks of inputs, or it
is non-const and cannot work with chunks of input. These functions were
none of these.

* Non-const processOne cannot be called in a SIMD-loop

* Make nosonar_node_api non-SIMD-only

With `sink::processOne` being non-const the merged graphs in this
benchmark are always non-SIMD. Consequently the processOne_simd calls
become ill-formed and need to be removed. Thus, there's no difference
anymore when compiling with DISABLE_SIMD and that extra benchmark
executable can go.

Signed-off-by: Matthias Kretz <m.kretz@gsi.de>